backend/requirements.txt
# LiveKit Agent
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-openai>=0.8.0
livekit-plugins-elevenlabs>=0.7.0
livekit-plugins-silero>=0.6.0

# RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.0
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
faiss-cpu>=1.7.4
pypdf>=3.17.0

# Groq - FREE LLM alternative!
groq>=0.4.0

# Utils
python-dotenv>=1.0.0
aiohttp>=3.9.0
livekit-api>=0.6.0

